Honor export_name in field specialization c-exposed exclusion
